refactor, main.js -> index.js and src/cli.js -> bin/abell.js
