+++ tests/index.spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const main = require('../src/main.js');
+const main = require('../src/index.js');
 
 describe('main()', () => {
   it('should export neccessary functions', () => {
